Fix cycle dependencies issue in UserInputsList
